Afuller/check assert (#21576)
### Ticket
N/A

### Problem description
ASan and TSan were missing TT_ASSERT because TT_ASSERT was guarded by
-DDEBUG which was being set later than when we initialized the flags for
ASan and TSan.

We are referencing CMAKE_<lang>_FLAGS_RELWITHDEBINFO here:
https://github.com/tenstorrent/tt-metal/blob/main/cmake/sanitizers.cmake#L24
We invoke that code on line 35:
https://github.com/tenstorrent/tt-metal/blob/main/CMakeLists.txt#L35
But we're not actually overriding it with the -DDEBUG which is expected
to exist for TT_ASSERT until line 89:
https://github.com/tenstorrent/tt-metal/blob/main/CMakeLists.txt#L89
This means we're using the original value and not the updated value.
Thus we missed out on -DDEBUG.

### What's changed
Ensure that custom RelWithDebInfo flags are set before we reference them
for setting ASan and TSan flags.


Confirmed functional:
https://github.com/tenstorrent/tt-metal/actions/runs/14801193838/job/41560984194?pr=21576#step:6:59