chore: update setup_testing_env.sh to display info in case of error (#170)
### Ticket
None

### Problem description
Currently, we are experiencing [random
failures](https://github.com/tenstorrent/tt-llk/actions/runs/14880439156/job/41787137193?pr=169)
caused by network issues when downloading the SFPI compiler. However,
the error is misleading as it only appears during the pytest execution
step. Instead, the failure should occur earlier, during the step where
the SFPI compiler is being downloaded.

### What's changed
With this change, GH pipeline should stop if it fails to fetch the
compiler or to unpack it. Pytests shouldn't even run.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update