ci: run LLK test infra conditionally, depending on the changed files (#171)
### Ticket
None

### Problem description
Currently, both Blackhole and Wormhole tests are being run regardless of
the changes introduced in the PR. Since we only have a limited number of
WH runners and just one BH runner available, optimizing resource usage
is crucial.

### What's changed
With this change, the CI pipeline will only run the appropriate tests.
For example:
- If Wormhole-specific code is changed, tests will only run on the
Wormhole runner.
- If Blackhole-specific code is changed, tests will only run on the
Blackhole runner.
- If unrelated files (e.g., README) are changed, unit tests will not
run, as they are not expected to fail.

This ensures efficient use of resources.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update