Fix dead stores (#21600)
### Ticket
N/A

### Problem description
Dead stores (storing a value in a variable but never reading it back)
can be a symptom of a logic error. Either forgetting to do an operation
with the variable, or doing the operation on the wrong variable.

The compiler will flag dead stores for us. But first we need to clean up
existing ones.

**TO REVIEWERS: Please audit the flagged code for logic errors involving
the variable in question. I suspect most are probably a result of
copy/paste or refactoring that didn't fully clean up afterwards. But
there's no guarantee of that. We should ensure each dead store was not a
result of a logic error.**

### What's changed
* Enabled the compiler warning (except in Release mode)
* Delete the dead-store vars

### Checklist
- [X] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/14812397053)