Make MemoryConfig a class with accessors (#21679)
### Ticket

### Problem description
We're working on support of ND-sharding in Metal & TTNN
This requires an extension of MemoryConfig, which is problematic because
it is a transparent struct, so any user can read/write sharding
parameters at any point.

### What's changed
Make MemoryConfig a class with accessors, allowing for future changes.
Mass refactoring to the API changes, all meaningful changes to
MemoryConfig will be done separately.
The end goal is to allow users to use both simple and ND sharding APIs,
while allowing to use OPs which are unaware of ND-sharding for cases
when its convertible to simple sharding and vice versa.

### Checklist
- [x] [All post commit CI
passes](https://github.com/tenstorrent/tt-metal/actions/runs/14891249012)
- [x] [Model regression CI
passes](https://github.com/tenstorrent/tt-metal/actions/runs/14854557641)
- [x] New/Existing tests provide coverage for changes