Ensure matching hooks on Buffer allocation/deallocation (#21806)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/21315

### Problem description
Currently there is an issue with inconsistent hooks on Buffer
allocation/deallocation, which happens specifically when a Buffer was
cached by an OP (stored in override runtime args callbacks) and
preserved by ProgramCache.
This PR ensures that program cache is cleared at the appropriate time of
device closing to ensure correct handling and adds diagnostic to easily
find such issues in the future.

We shouldn't be storing the Buffers inside of override runtime arguments
to begin with, because it's basically a memory leak with program cache
enabled, but this will be addressed separately for each of the OPs
individually.

### What's changed
Moved clearing program cache up in the Device::close()
Stored a set of hooked buffers in GraphTracker with validation that we
hook deallocation for exactly the same buffers where we hooked the
allocation
Stored in buffer whether its allocation was hooked and avoiding
deallocation for them even if the hook was removed

### Checklist
- [x] [All post commit CI
passes](https://github.com/tenstorrent/tt-metal/actions/runs/14897246210)
- [x] New/Existing tests provide coverage for changes