`DistributedTensorConfig` is the abstraction responsible for describing how a tensor is moved from host to device. It is currently only attached to "multi device host" storage variant. In practice, single-buffer host-side storage (currently `HostStorage`) behaves like a `Tensor` fully replicated across mesh.
The change to unify the APIs is needed to subsequently unify the "multi
device host" storage with regular "host" storage. It will be further
extended to be multi-host aware.
Align reality with the code by providing an unconditional getter for
`DistributedTensorConfig` on the `Tensor` / `TensorAttributes`.

Align single-device host-side constructors of `Tensor`, to accept
`HostBuffer` as "physical data" representation. This avoids misuse and
reliance on `Storage` classes, which are internal and not intended for
external use.
- [X] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/14909202060)
- [X] [T3K
tests](https://github.com/tenstorrent/tt-metal/actions/runs/14898442006)
- [X] New/Existing tests provide coverage for changes