#19609: Update unary backward ops  (#21353)
### Ticket
Link to Github Issue #19609 

### Problem description
Tests failing on backward exp, elu, logy after adding seeding.

### What's changed

- backward_exp: Updated the logic by removing special handling for inf,
which was needed for GS but not for WH. This operation uses the forward
exp, which supports a limited input range, so the input range was
adjusted to avoid issues with large values.

- backward_elu: Fixed the logic by changing the condition from "greater
than or equal to zero" to "greater than zero", which correctly resolves
the issue.

- backward_xlogy: Since this operation uses log, which doesn't support
negative values, the input range was updated to prevent errors.

### Checklist
- [ ] [All post commit
cI](https://github.com/tenstorrent/tt-metal/actions/runs/14747153135)