feat: paralelize test execution (#182)
### Ticket
None

### Problem description
Our tests currently take more than 30 minutes to complete. To address
this, we can reduce the overall runtime by splitting the tests into
groups and running them on different runners in parallel. While this
approach won't provide a full 5x speedup due to the overhead of
downloading Docker images, it should still result in a noticeable
improvement.

### What's changed
1. **Wormhole tests**: These tests now run on 5 runners in parallel,
reducing the overall runtime.
2. **Blackhole tests**: These tests continue to run on a single runner
for now, as we currently have only one available. As more runners become
available in the future, we will increase the number of parallel runs
accordingly.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update