fix: revert "Improve reciprocal precision. (#155)" (#183)
### Ticket
None

### Problem description
We have to revert #155, because it's causing issues in tt-metal's ttnn
tests. The tests seem off, but this needs to be investigated further,
and we're blocked by this commit. So, we will revisit this change after
some high priority changes land into metal.

### What's changed
This reverts commit 39f3826ca8a7befef49fdf8425789c04aea7b5c5.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update