#19177: Unconditionally attach `DistributedTensorConfig` to `TensorAttributes` (#21905)
### Ticket
#19177

### Problem description
Rollforward https://github.com/tenstorrent/tt-metal/pull/21813, reverted
in
https://github.com/tenstorrent/tt-metal/commit/0718236a4f75cae8fff42ceeb3584613c71ddb9d
due to a test failure in
`tests/ttnn/unit_tests/operations/reduce/test_cumsum.py`.

### What's changed
The problem was that conversion to dtype didn't convert back to tile
layout. This was omitted in the original PR and re-added in the
rollforward.

The fix is added in
https://github.com/tenstorrent/tt-metal/pull/21905/commits/535ed56a2de783ebedba66ddc772a19dc7018a63

### Checklist
- [X] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/14933322133)
- [X] New/Existing tests provide coverage for changes