fix: remove hardware-specific header files shared with tt-metal (#189)
### Ticket
#177

### Problem description
Currently, we have several header files that reside in the tt-metal
repository but are required for the LLK test infrastructure to compile
and execute. These files primarily contain definitions (e.g. memory
sizes, register addresses, etc.). Until now, copies of these files were
stored in the tt-llk repository to break the dependency. However, since
they are frequently updated in the tt-metal repository, we need to
continuously track these changes to keep them in sync. This approach is
not sustainable.

### What's changed
The proposed solution is to download these files from tt-metal
dynamically whenever LLK tests are run. This approach is minimally
intrusive, as it only affects our repository and does not impact others.
The change in this PR does just that: it downloades the required files
on the fly, just in time for tests to run.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update