Fix assert syntax (#21932)
### Ticket
N/A

### Problem description
These asserts will never fail.
In APC we are told that fact 104 times and that gets noisy in the
workflow summary 🙃

### What's changed
Fixed the syntax

AFM: I modified
`test_repeat.py::test_pc_with_different_shapes_in_sequence` to limit its
scope to only testing for the caching behavior of `ttnn.repeat`.
Previously, in addition to the incorrect `assert` syntax, it made
assumptions that the caching behavior of a handful of ops was similar to
`ttnn.repeat`. The op flow of that test, which intended to reproduce a
bug ticket, remains.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/14982030459)

---------

Co-authored-by: Adrian F. Morrison <amorrison@tenstorrent.com>