Assert on page table size in simple_text_demo to avoid OOB array read (#22014)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/21971

### Problem description
When increasing `max_generated_tokens` without increasing the size of
the paged cache we hit hangs in WH and BH due to OOB array reads in
`paged_update_cache`.

### What's changed
Assert in the demo that the paged cache is large enough to manage the
largest prefill user and the number of tokens generated.