Update paged_fill_cache to support tensor batch_idxs input  (#22029)
### Problem description
We need tensor batch_idxs input for tracing. If a trace is capture for
an integer argument, it is executed for the same argument, but if it is
ttnn tensor which is read from memory, we can override the batch_id at
trace execution.

### What's changed
*  support tensor batch_idxs input 
*  skip fill cache for block when page_id = -1 

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/14999777761)
CI passes