#22090: Clean up `Tensor` getters for various attributes (#22092)
### Ticket
#22090

### Problem description
The `get_<x>` were used in async runtime, and are no longer needed.

### What's changed
Removing all of the legacy getters in one go is going to break clients,
so the first step is to mark them as deprectated.

Other minor change:
* Added overloads for `logical_volume` and `padded_volume` (as opposed
to just one "volume" which is ambiguous). Used `uint64_t`, since this is
the actual type returned by `Shape::volume`. It is not inconceivable
that we can get overflows with uint32_t.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/15026608572)
- [x] New/Existing tests provide coverage for changes