ci: prevent conflicts when fetching external forks (#201)
### Ticket
None

### Problem description
This PR should be fixing errors like this:
```
PR is from a fork: jasondavies/tt-llk
Adding contributor's fork as remote...
error: remote fork already exists.
Error: Process completed with exit code 3.
```

### What's changed
The way we are fetching forks is changed, so we will gracefully continue
if the fork is already present.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update