Minimise data movement in transpose_dest. (#176)
### Ticket

#175

### Problem description

The original code only copied data between B and Dst. When swapping the
middle two tiles, this meant using rows 0-15 of B as a temporary buffer
for one of the middle tiles before later copying back to Dst and then
back to B to rows 16-31 for transpose, and finally to the free tile
position in Dst.

Note that MOVA2D supports 8-row copies, so that only two instructions
are required to copy a 16x16 tile, rather than 4 instructions as
required for the other MOV* instructions that only support up to 4 rows
at a time. Therefore it makes more sense to use A as the temporary
buffer.

This also opens up the possibility of reducing the total number of tile
copies, since we also have MOVB2A, and we can exploit this to do the
following for the 2nd and 3rd tiles:

1. Copy 2nd tile to B
2. Transpose B
3. Copy to A, which frees up the original 2nd tile in Dst.
4. Copy the 3rd tile to B
5. Transpose B
6. Overwrite 2nd tile in Dst
7. Finally, copy the transposed tile from A to Dst, overwriting the 3rd
tile.

This reduces the total number of _tile movements_ from 10 to 9.

### What's changed

- Replace the use of rows 0-15 of SrcB with SrcA
- 0x3ff mask added to negative destination offsets to avoid strange
macro expansion effects when using negative numbers there
- MOP template 0 used to store 7 instructions that reference a 15-slot
replay buffer

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

### Checklist
<!-- These are required steps and need to be run from tt-metal
repository's Actions. Use links below and replace them with your run -->
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)

---------

Co-authored-by: Nikola Velickovic <nvelickovic@tenstorrent.com>
Co-authored-by: Filip Vranic <fvranic@tenstorrent.com>