SDXL optional VAE host fallback (#22137)
### What's changed
Added an optional fallback for group norm in resnet blocks. The pcc
issue of the op seems to be resolved but when using the device version
of the op the overall pcc of the VAE decoder drops to 0.85 and 0.89 for
the wrapper. When using the no fallback implementation to generate an
image we get visually equivalent image to the one with fallback. But in
the absence of accuracy testing this is not a good enough metric to
completely remove fallback option.

Because of this we keep the option to fallback and test decoder and
wrapper in CI in both forms. This is done to avoid masking a potential
regression in pcc by the overall lower score while minimising the number
of additional tests in CI.

Generated image:

![output](https://github.com/user-attachments/assets/f7147b05-d9a2-4483-a7d4-1498200185c3)


### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/15041275344)
CI passes
- [x] [(Single-card) Frequent model and ttnn
tests](https://github.com/tenstorrent/tt-metal/actions/runs/15132014156/job/42539947230)
CI passes