Simplify querying of dispatch related cores (#22043)
### Ticket
#18726

### Problem description
- Simplify device init reset/close. For FD on Fabric, we need to be able
to stall on dispatch, then stall on mux, and finally terminate fabric.
- It's difficult to do this right now with the current implementation
because dispatch core manager can query by servicing device id but not
which device the core has been allocated on.

### What's changed
- Topology to register which cores have been used for dispatch or
routing
- Simplify device termination.
- Device level: assert all worker / non dispatch / routing cores to be
done.
- Device pool level: Terminate dispatch kernels. Stall for all dispatch
cores on the MMIO chip to be done. and then terminate the fabric
routers. Once the MMIO chip cores are done it means the remote ones are
done as well because at the end of the dispatch kernels it syncs for
other dispatch kernels to return all credits.
- Remove get_associated_dispatch_virtual_cores from Device
- Moved `is_dispatch_firmware_active_` from Device to DevicePool

### Checklist
APC
https://github.com/tenstorrent/tt-metal/actions/runs/15142164071
BH
https://github.com/tenstorrent/tt-metal/actions/runs/15142169958
T3k
https://github.com/tenstorrent/tt-metal/actions/runs/15052398907
https://github.com/tenstorrent/tt-metal/actions/runs/15077426660
TG
https://github.com/tenstorrent/tt-metal/actions/runs/15073424560