Activating Data Format Inference model with Input Output format parametrization Flexibility for same formats  (#222)
### Problem description
<!-- Provide context for the problem. -->
User should run all tests using input-output parametrization in order to
utilize data format inference model. This is very useful because all
input-output format combinations are supported by the data format
inference model. If we hard code the combinations we might make a
mistake as some hard-coded combinations with certain flags such as
dest_Acc are not supported on certain architecture, and the data format
inference model takes care of this in order to support all input-output
combinations and increase coverage.

Fused tests that incorporate 2 or more llk pipeline runs currently don't
utilize data format inference model because all the formats are the same
(in order to reuse output of first call as input to second llk pipeline
run ...)
However if all formats are the same this isn't always supported by
hardware for example if all formats are float16_an and dest is in
float32 mode, on bh this works but fails on wh. The data format
inference model accounts for this and sets the pack input format to
float32.

By running on all tests with input-output format we activate the data
format inference model to create combinations for each input-output
combination. This increases coverage and we don't have to skip test
cases that work with minor tweaking.

### What's changed
<!-- Describe the approach used to solve the problem.
Summarize the changes made and its impact. -->
Added flag to input-output formats sweep when we want input and output
format to be the same.
Run tests using this flag when input and output need to be the same so
that the output is reused for fused tests.
### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update