Register termination info in topology (#22493)
### Ticket
[#18726](https://github.com/tenstorrent/tt-metal/issues/18726)

### Problem description
- When running Dispatch on Fabric we need to terminate Dispatch ->
Fabric Mux -> Fabric
- If we get the dispatch kernel to send the termination signal, it could
terminate the MUX before other dispatch clients are done.
- Another option is to assign a main dispatch kernel which will
terminate the MUX on that device, but that would need to sync with all
other dispatch cores to be done on the device which is more complicated
than syncing on the host

### What's changed
- Kernels created in topology to register their termination information
using `GetTerminationInfo`. DevicePool::close_devices will send
termination info to each core that was registered
- Minor cleanup: Put FDKernels into tt::tt_metal namespace

```cpp
struct TerminationInfo {
    CoreCoord logical_core;  // Logical core coordination
    CoreType core_type;      // Core Type
    uint32_t address;        // Termination signal address in L1
    uint32_t val;            // Termination signal value
};
```

### Checklist
APC
https://github.com/tenstorrent/tt-metal/actions/runs/15198863601