fix: detect chip architecture automatically (#232)
### Ticket
https://github.com/tenstorrent/tt-llk/issues/224

### Problem description
There was a regression in behavior wrt automatic chip detection.
In case external testing environment is used (not from the tt-llk docker
image), venv wasn't automatically sourced.

### What's changed
This issue with chip detection should be resolved now. Logic was
misplaced.
venv sourcing issue caused some confusion, because it wasn't mentioned
in the testing readme file. This is also addressed.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update