Fix kernel_text_offset alignment
kernel_text_offset isn't 4-byte aligned, so the compiler generates multiple
instructions to load it.

Move around pad1 to ensure kernel_text_offset is aligned. Also add
static_asserts to catch these kinds of errors in the future. They also caught
an issue with host_assigned_id (which is only used on the device when tracy is
enabled), so fix that as well.