Fix missing dependency (#22548)
### Ticket
N/A

### Problem description
TT-STL is missing a dependency on span. It only accidentally works if
Boost is installed on the system and is a sufficient version to include
span.
TT-STL's test is missing a dependency on UUT. It only accidentally works
because include dirs were being assumed. And even then only if Boost is
installed on the system and is a sufficient version to include span.

### What's changed
Fixed the dependency logic so that it works even if deps come in via
CPM.