T3K vLLM online test (#22359)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/21651

### Problem description
vLLM often has to catch-up as the models progress.
We want to couple vLLM and models more strongly.

### What's changed
Add a T3K test for vLLM in online mode.
It should be added to Package and Release.

### Checklist
This is a new test, it's not part of any test suite yet.
The test is successful on Metal CI ✅
https://github.com/tenstorrent/tt-metal/actions/runs/15146526002