Fix slotmap default operator<=> return type
Default three-way comparison returns std::strong_ordering
(or weak_ordering or partial_ordering), not convertible to bool,
so bool is a bug here; this declaration fail on instantiation.

This PR also changes the operator to a non-member friend function
(a 'hidden friend' as it's defined in-class) which allows the
parameters to be passed by-value rather than by-reference -
assuming that T is a trivial type this can give better codegen.

(This is a first test PR, with no ticket, to test the process.)