Exclude ci-stress-1 in demo tests (#22633)
### Ticket
Link to Github Issue

### Problem description
`performance-ci-stress-1` was added in
https://github.com/tenstorrent/tt-metal/commit/ef9628d4340e9304036721df6ceed3ed1163834f
to enable stress testing llama8b on blackhole.
However, we shouldn't run it as part of the single-card and t3k demo
tests (takes 30 mins with 50k iterations, we are also missing the tensor
caches for the test params which is causing it to fail in CI)

### What's changed
Exclude `performance-ci-stress-1` from single card and t3k demos.

### Checklist
- [ ] Single card demos:
https://github.com/tenstorrent/tt-metal/actions/runs/15277820712
- [ ] T3k demos:
https://github.com/tenstorrent/tt-metal/actions/runs/15283508872