ci: add pull request write permissions to workflows (#258)
### Ticket
None

### Problem description
After we moved to GH Enterprise, our comment posting job started failing
in all pull requests.

### What's changed
This change elevates permissions of workflows running on PRs to write,
so it should fix our problem with comment posting job. It should now be
able to write comments on PRs.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update