[skip ci] Do not attempt to load tt-fabric tests.  Presumed static object badness inside. (#22470)
### Ticket
N/A

### Problem description
Sometimes the binary hangs when attempting to merely list the tests
inside. eg:
https://github.com/tenstorrent/tt-metal/actions/runs/15175175128/job/42674836336#step:10:1148
Presumably there is some static object that is causing grief, otherwise
that should never happen. Unfortunately it doesn't always happen, so
difficult to root cause.

### What's changed
Do not probe for the individual tests inside the binary. I'd be
surprised if anybody is using that feature anyway.