fix: speed up test build times (#265)
### Ticket
None

### Problem description
A large portion of the test execution time was actually spent on build
time. We were building ELF files inefficiently, and there was room for
improvement by parallelizing the make process. This should have
significant impact on the amount of time we spend on running the tests.

### What's changed
We now run make in parallel using 6 cores instead of building
sequentially on a single core. I chose this number after testing locally
to measure the improvement. The tests were performed on a machine with
12 cores, but using all available cores actually resulted in slower
builds compared to using half of them. After several tests with
different core counts, I found that 6 cores consistently gave the best
results, achieving around a 3x speed improvement.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update