#22205: Implement on host data preparation for ND sharding for future perf improvements (#22847)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/22205

### Problem description
We need to optimize Tensor transfer to/from device for ND sharding. This
can be done by reordering data on host to avoid a lot of small
individual data transfers.

### Approaches to optimization
1. Reorder data on host to minimize amount of the data transfers to
device each time a Tensor is being written
2. Reorder data on host when a Tensor is being created, similarly to how
we tilize Tensor on host
3. Write the Tensor data as-is and use on-device resharding

The approach 3 is a bit problematic, because resharding can't be done in
place, so it would require extra memory which might not be available.
Currently we're investigating approaches 1 and 2, and will do a proper
profiling and perf analysis to determine our next steps.

### What's changed
Implemented function to pack/unpack ND sharded data, ordering all data
which go on each core sequentially

### Checklist
- [x] [All post commit CI
passes](https://github.com/tenstorrent/tt-metal/actions/runs/15383721220)
- [x] New/Existing tests provide coverage for changes